requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-autotranslate==1.1.0
django-countries==5.5
django-crispy-forms==1.8.0
django-decouple==2.1
django-filebrowser-no-grappelli==3.7.8
django-modeltranslation==0.13.4
django-rest-framework==0.1.0
django-rosetta==0.9.3
django-tinymce4-lite==1.7.5
djangorestframework==3.10.3
futures==3.1.1
goslate==1.5.1
idna==2.8
jsmin==2.2.2
Pillow==6.2.1
polib==1.1.0
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
